[CIR][AArch64] Upstream vset_lane_*/vsetq_lane_* NEON builtins, Part of #185382
